Added a "TriggerMessage" class which is a wrapper for the UDP trigger message (defined in the TriggerPacket struct). Updated TriggeredFragmentGenerator and EventStore to use this class.
EventStore will populate the trigger message timestamp field with the timestamp of the fragment that opened the event. Unless significant processing is happening in a BoardReader, this should be close to the desired value. Care should be taken to tune the offset and window parameters accordingly.

TriggeredFragmentGenerator will, when running in Window mode, wait until it has a data point with a timestamp greater than the trigger's timestamp before returning data. This ensures that al
l possible data that satisfy the trigger are sent.