Adding RoutingMaster functionality.
Copied most of the actual application code from Aggregator and
BoardReader to make RoutingMaster{Main,App,Core}.

Added new artdaq plugin type, the RoutingMasterPolicy, to allow users to
create custom rules for sequencing event builders. Example
RoundRobin_policy provided.

RoutingPacket.hh contains the three types of network packets I expect to
be sending back and forth: RoutingPacket contains the current table
information and is sent to BoardReaders. RoutingPacketAck is sent by the
BoardReader to acknowledge and validate reception of the RoutingPacket.
RoutingToken is sent from the EventStore in the EventBuilder whenever an
event is completed.

None of the actual socket code is implemented yet, this is all
in-progress.