Added "sleep_on_no_data_us" configuration parameter, which determines if getDataLoop will sleep after calling getNext_ when no data is returned. Default is no sleep.
Added metric "Avg Data Acquisition Time" to getDataLoop which record the
amount of time spent in getNext_. (When not running with the separate
data thread, this time is recorded by the "Avg Input Wait Time" metric).
"Avg Input Wait Time" in data thread mode records the amount of time that
applyRequests does work.