Add BrokenTransferTest and broken_transfer_driver, which are set up to test the following situations: 1. Sender pauses for a non-trivial amount of time 2. Receiver pauses for a non-trivial amount of time 3. Sender is killed and reconnects 4. Receiver is killed and reconnects
Right now, the test is implemented as two senders sending to one
receiver, due to limitations in TCPSocketTransfer (can't have more than
one receiving rank per process). There is also no routing involvement,
even to a fake destination, so the hang mentioned in #23050 is not
currently exposed by this test.