Feature #23131
I noticed that there was a potential logic flaw in the handling of
missing Fragment ID tracking. If the first event has Fragment ID 1 only,
then the second has Fragment ID 2 only, followed by a third with both ID
1 and 2, neither the first nor the second would be flagged as missing
Fragments. This commit removes the "expected_fragments_per_event_"
member, instead relying on the size of the expected_fragmentID_list_,
which is always updated as part of the missing Fragment ID check.